fix: ground_seg is now separate from occupancy_cpp
