+++ launch_node_definitions.py
@@ -52,7 +52,7 @@ grid_summation = Node(
 )
 
 ground_seg = Node(
-    package='occupancy_cpp',
+    package='ground_seg',
     executable='ground_segmentation_exe',
     parameters=[
         {'sensitivity': 0.13},
